Move config type to types instead of random component folder
