+++ src/cards/mediocre-media-player-card-editor.tsx
@@ -1,5 +1,5 @@
 import { MediocreMediaPlayerCardEditor } from "../components";
-import { MediocreMediaPlayerCardConfig } from "../components/MediaPlayerCommon";
+import { MediocreMediaPlayerCardConfig } from "../types";
 import { CardEditorWrapper } from "../utils";
 
 class MediocreMediaPlayerCardEditorWrapper extends CardEditorWrapper<MediocreMediaPlayerCardConfig> {
