--- src/components/MediaPlayerCommon/index.ts
@@ -1,1 +1,0 @@
-export * from "./config";
