+++ src/components/MediocreMassiveMediaPlayerCard/MediocreMassiveMediaPlayerCardEditor.tsx
@@ -1,5 +1,5 @@
 import { HomeAssistant } from "custom-card-helpers";
-import { MediocreMassiveMediaPlayerCardConfig } from "../MediaPlayerCommon/config";
+import { MediocreMassiveMediaPlayerCardConfig } from "../../types/config";
 import { useCallback } from "preact/hooks";
 import {
   Button,
