+++ src/components/MediocreMassiveMediaPlayerCard/components/Track.tsx
@@ -2,7 +2,7 @@ import { useContext, useMemo } from "preact/hooks";
 import styled from "@emotion/styled";
 import { ProgressBar } from "../../ProgressBar";
 import { CardContext, CardContextType } from "../../../utils";
-import { MediocreMassiveMediaPlayerCardConfig } from "../../MediaPlayerCommon";
+import { MediocreMassiveMediaPlayerCardConfig } from "../../../types";
 
 const TimeWrap = styled.div`
   display: flex;
