+++ src/components/MediocreMediaPlayerCard/components/PlaybackControls.tsx
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "preact/hooks";
 import styled from "@emotion/styled";
 import { IconButton } from "../../IconButton";
 import { CardContext, CardContextType } from "../../../utils";
-import { MediocreMediaPlayerCardConfig } from "../../MediaPlayerCommon/config";
+import { MediocreMediaPlayerCardConfig } from "../../../types/config";
 import { useSupportedFeatures } from "../../../hooks/useSupportedFeatures";
 
 const PlaybackControlsWrap = styled.div`
@@ -15,7 +15,7 @@ const PlaybackControlsWrap = styled.div`
   margin-left: -4px; // compensate for icon button padding
 `;
 
-const ControlButton = styled(IconButton)<{ muted?: boolean }>`
+const ControlButton = styled(IconButton) <{ muted?: boolean }>`
   opacity: ${props => (props.muted ? 0.8 : 1)}; // reduce opacity if muted
 `;
