+++ src/components/MediocreMediaPlayerCard/components/PlayerInfo.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useContext } from "preact/hooks";
 import { CardContext, CardContextType } from "../../../utils";
-import { MediocreMediaPlayerCardConfig } from "../../MediaPlayerCommon/config";
+import { MediocreMediaPlayerCardConfig } from "../../../types/config";
 import styled from "@emotion/styled";
 import { IconButton } from "../../IconButton";
 import { fireEvent } from "custom-card-helpers";
