+++ src/components/MediocreMediaPlayerCard/components/SpeakerGrouping.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useContext, useMemo, useState } from "preact/hooks";
 import styled from "@emotion/styled";
-import type { MediocreMediaPlayerCardConfig } from "../../MediaPlayerCommon/config";
+import type { MediocreMediaPlayerCardConfig } from "../../../types/config";
 import { Slider } from "../../Slider";
 import { IconButton } from "../../IconButton";
 import { CardContext, CardContextType } from "../../../utils";
