+++ src/components/MediocreMediaPlayerCard/components/VolumeSlider.tsx
@@ -2,7 +2,7 @@ import { useCallback, useContext, useMemo } from "preact/hooks";
 import styled from "@emotion/styled";
 import { IconButton } from "../../IconButton";
 import { CardContext, CardContextType } from "../../../utils";
-import { MediocreMediaPlayerCardConfig } from "../../MediaPlayerCommon/config";
+import { MediocreMediaPlayerCardConfig } from "../../../types/config";
 import { Slider } from "../../Slider";
 import { Fragment } from "preact/jsx-runtime";
 
@@ -15,7 +15,7 @@ const VolumeContainer = styled.div`
   margin-top: auto;
 `;
 
-const ControlButton = styled(IconButton)<{ muted?: boolean }>`
+const ControlButton = styled(IconButton) <{ muted?: boolean }>`
   opacity: ${props => (props.muted ? 0.8 : 1)}; // reduce opacity if muted
 `;
