+++ src/components/index.ts
@@ -2,7 +2,6 @@ export * from "./Chip";
 export * from "./FormElements";
 export * from "./Icon";
 export * from "./IconButton";
-export * from "./MediaPlayerCommon";
 export * from "./MediocreChipMediaPlayerGroupCard";
 export * from "./MediocreMassiveMediaPlayerCard";
 export * from "./MediocreMediaPlayerCard";
