+++ src/types/index.ts
@@ -1,2 +1,3 @@
-export * from "./actionTypes";
-export * from "./ha";
+export * from './actionTypes';
+export * from './config';
+export * from './ha';
